with-supertokens: reduced bundle size by removing node lib from bundle
Linting fix
