+++ examples/with-supertokens/config/backendConfig.js
@@ -1,22 +1,6 @@
 import ThirdPartyEmailPasswordNode from 'supertokens-node/recipe/thirdpartyemailpassword'
 import SessionNode from 'supertokens-node/recipe/session'
-
-import ThirdPartyEmailPasswordReact from 'supertokens-auth-react/recipe/thirdpartyemailpassword'
-import SessionReact from 'supertokens-auth-react/recipe/session'
-
-const port = process.env.APP_PORT || 3000
-const websiteDomain =
-  process.env.APP_URL ||
-  process.env.NEXT_PUBLIC_APP_URL ||
-  `http://localhost:${port}`
-const apiBasePath = '/api/auth/'
-
-let appInfo = {
-  appName: 'SuperTokens Demo App',
-  websiteDomain,
-  apiDomain: websiteDomain,
-  apiBasePath,
-}
+import { appInfo } from './appInfo'
 
 export let backendConfig = () => {
   return {
@@ -48,25 +32,5 @@ export let backendConfig = () => {
       SessionNode.init(),
     ],
     isInServerlessEnv: true,
-  }
-}
-
-export let frontendConfig = () => {
-  return {
-    appInfo,
-    recipeList: [
-      ThirdPartyEmailPasswordReact.init({
-        emailVerificationFeature: {
-          mode: 'REQUIRED',
-        },
-        signInAndUpFeature: {
-          providers: [
-            ThirdPartyEmailPasswordReact.Google.init(),
-            ThirdPartyEmailPasswordReact.Github.init(),
-          ],
-        },
-      }),
-      SessionReact.init(),
-    ],
   }
 }
