+++ examples/with-supertokens/config/frontendConfig.js
@@ -1,0 +1,23 @@
+import ThirdPartyEmailPasswordReact from 'supertokens-auth-react/recipe/thirdpartyemailpassword'
+import SessionReact from 'supertokens-auth-react/recipe/session'
+import { appInfo } from './appInfo'
+
+export let frontendConfig = () => {
+  return {
+    appInfo,
+    recipeList: [
+      ThirdPartyEmailPasswordReact.init({
+        emailVerificationFeature: {
+          mode: 'REQUIRED',
+        },
+        signInAndUpFeature: {
+          providers: [
+            ThirdPartyEmailPasswordReact.Google.init(),
+            ThirdPartyEmailPasswordReact.Github.init(),
+          ],
+        },
+      }),
+      SessionReact.init(),
+    ],
+  }
+}
