+++ examples/with-supertokens/pages/api/user.js
@@ -1,9 +1,9 @@
 import { superTokensNextWrapper } from 'supertokens-node/nextjs'
 import { verifySession } from 'supertokens-node/recipe/session/framework/express'
 import supertokens from 'supertokens-node'
-import * as SuperTokensConfig from '../../config/supertokensConfig'
+import { backendConfig } from '../../config/backendConfig'
 
-supertokens.init(SuperTokensConfig.backendConfig())
+supertokens.init(backendConfig())
 
 export default async function user(req, res) {
   await superTokensNextWrapper(
